gcf1/requirements.txt
requests==2.25.1
google-auth==1.1.0
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.12.5
chardet==4.0.0
idna==2.10
requests==2.25.1
soupsieve==2.2.1
urllib3==1.26.4
v==0.0.0
pyodbc==4.0.30
pandas==1.2.4
numpy==1.20.3
mysql-connector-python==8.0.25
protobuf==3.17.2